Use wait_for_transaction util everywhere
